requirements.txt
# transformers 
huggingface_hub
# torch
# pillow
# PyPDF2
python-dotenv
# smolagents